+++ Drop-In/src/test/java/com/braintreepayments/api/DropInActivityUnitTest.java
@@ -62,7 +62,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onCreate_whenAuthorizationIsInvalid_finishesWithError() {
         String authorization = "not a tokenization key";
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = mock(DropInClient.class);
         when(dropInClient.getAuthorization()).thenReturn(mock(InvalidAuthorization.class));
@@ -80,7 +80,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onResume_deliversBrowserSwitchResult() {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = mock(DropInClient.class);
 
@@ -93,7 +93,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onResume_whenBrowserSwitchResultExists_finishesActivity() {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInResult result = mock(DropInResult.class);
         PaymentMethodNonce nonce = mock(PaymentMethodNonce.class);
@@ -113,7 +113,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onResume_whenBrowserSwitchReturnsUserCanceledException_setsUserCanceledErrorInViewModel() {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         Exception error = new UserCanceledException("User canceled 3DS.");
         DropInClient dropInClient = new MockDropInClientBuilder()
@@ -132,7 +132,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onResume_whenBrowserSwitchError_forwardsError() {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         Exception error = new Exception("A 3DS error");
         DropInClient dropInClient = new MockDropInClientBuilder()
@@ -151,7 +151,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onActivityResult_whenDropInResultExists_finishesActivity() {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInResult result = mock(DropInResult.class);
         PaymentMethodNonce nonce = mock(PaymentMethodNonce.class);
@@ -171,7 +171,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onActivityResult_whenResultIsUserCanceledException_setsUserCanceledErrorInViewModel() {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         Exception error = new UserCanceledException("User canceled 3DS.");
         DropInClient dropInClient = new MockDropInClientBuilder()
@@ -190,7 +190,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onActivityResult_whenError_handlesError() {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         Exception error = new Exception("A 3DS error");
         DropInClient dropInClient = new MockDropInClientBuilder()
@@ -209,7 +209,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onActivityResult_forwardsResultToDropInClient() {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = new MockDropInClientBuilder()
                 .build();
@@ -226,7 +226,7 @@ public class DropInActivityUnitTest {
     @Test
     public void setsIntegrationTypeToDropinForDropinActivity() {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
         setupDropInActivity(authorization, mock(DropInClient.class), dropInRequest, "sessionId");
         mActivityController.setup();
 
@@ -238,7 +238,7 @@ public class DropInActivityUnitTest {
     @Test
     public void sendsAnalyticsEventWhenShown() {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
         setupDropInActivity(authorization, mock(DropInClient.class), dropInRequest, "sessionId");
         mActivityController.setup();
 
@@ -248,6 +248,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onVaultedPaymentMethodSelected_reloadsPaymentMethodsIfThreeDSecureVerificationFails() throws JSONException {
         DropInClient dropInClient = new MockDropInClientBuilder()
+                .authorization(Authorization.fromString(base64EncodedClientTokenFromFixture(Fixtures.CLIENT_TOKEN)))
                 .threeDSecureError(new Exception("three d secure failure"))
                 .shouldPerformThreeDSecureVerification(true)
                 .build();
@@ -256,7 +257,6 @@ public class DropInActivityUnitTest {
         threeDSecureRequest.setAmount("1.00");
 
         DropInRequest dropInRequest = new DropInRequest()
-                .clientToken(base64EncodedClientTokenFromFixture(Fixtures.CLIENT_TOKEN))
                 .threeDSecureRequest(threeDSecureRequest)
                 .requestThreeDSecureVerification(true);
 
@@ -276,7 +276,7 @@ public class DropInActivityUnitTest {
     @Test
     public void pressingBackExitsActivityWithResultCanceled() {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
         setupDropInActivity(authorization, mock(DropInClient.class), dropInRequest, "sessionId");
         mActivityController.setup();
 
@@ -289,7 +289,7 @@ public class DropInActivityUnitTest {
     @Test
     public void pressingBackSendsAnalyticsEvent() {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = new MockDropInClientBuilder()
                 .build();
@@ -304,7 +304,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onVaultedPaymentMethodSelected_whenShouldNotRequestThreeDSecureVerification_returnsANonce() throws JSONException {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = new MockDropInClientBuilder()
                 .shouldPerformThreeDSecureVerification(false)
@@ -328,7 +328,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onVaultedPaymentSelected_requestsThreeDSecureVerificationForCardWhenEnabled() throws Exception {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = new MockDropInClientBuilder()
                 .shouldPerformThreeDSecureVerification(true)
@@ -346,7 +346,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onVaultedPaymentMethodSelected_sendsAnAnalyticsEvent() throws JSONException {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = new MockDropInClientBuilder()
                 .collectDeviceDataSuccess("device data")
@@ -369,7 +369,7 @@ public class DropInActivityUnitTest {
                 .build();
 
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
         setupDropInActivity(authorization, dropInClient, dropInRequest, "sessionId");
         mActivityController.setup();
 
@@ -388,7 +388,6 @@ public class DropInActivityUnitTest {
     @Test
     public void onVaultedPaymentMethodSelected_returnsDeviceData() throws JSONException {
         DropInRequest dropInRequest = new DropInRequest()
-                .tokenizationKey(TOKENIZATION_KEY)
                 .collectDeviceData(true);
 
         String authorization = Fixtures.TOKENIZATION_KEY;
@@ -417,7 +416,7 @@ public class DropInActivityUnitTest {
     @Test
     public void selectingAVaultedPaymentMethod_returnsANonce() throws JSONException {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         List<PaymentMethodNonce> nonces = new ArrayList<>();
         PaymentMethodNonce paymentMethodNonce = PayPalAccountNonce.fromJSON(new JSONObject(Fixtures.PAYPAL_ACCOUNT_JSON));
@@ -446,7 +445,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onVaultedPaymentMethodSelected_whenCard_sendsAnalyticEvent() throws JSONException {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = new MockDropInClientBuilder()
                 .build();
@@ -463,7 +462,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onVaultedPaymentMethodSelected_whenPayPal_doesNotSendAnalyticEvent() throws JSONException {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = new MockDropInClientBuilder()
                 .build();
@@ -528,7 +527,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onSupportedPaymentMethodSelected_withTypePayPal_tokenizesPayPal() throws JSONException {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = new MockDropInClientBuilder()
                 .authorization(Authorization.fromString(authorization))
@@ -548,7 +547,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onSupportedPaymentMethodSelected_withTypeVenmo_tokenizesVenmo() throws JSONException {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = new MockDropInClientBuilder()
                 .authorization(Authorization.fromString(authorization))
@@ -568,7 +567,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onSupportedPaymentMethodSelected_withTypeGooglePay_requestsGooglePay() throws JSONException {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = new MockDropInClientBuilder()
                 .authorization(Authorization.fromString(authorization))
@@ -588,7 +587,7 @@ public class DropInActivityUnitTest {
     @Test
     public void onSupportedPaymentMethodSelected_withTypeUnknown_showsAddCardFragment() throws JSONException {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = new MockDropInClientBuilder()
                 .authorization(Authorization.fromString(authorization))
@@ -613,7 +612,7 @@ public class DropInActivityUnitTest {
 
     private void assertExceptionIsReturned(String analyticsEvent, Exception exception) {
         String authorization = Fixtures.TOKENIZATION_KEY;
-        DropInRequest dropInRequest = new DropInRequest().tokenizationKey(authorization);
+        DropInRequest dropInRequest = new DropInRequest();
 
         DropInClient dropInClient = new MockDropInClientBuilder()
                 .build();