+++ CHANGELOG.md
@@ -29,6 +29,7 @@
   * Remove `EditCardView`
   * Change `DropInPaymentMethodType#getCardTypes()` parameter from `Set` to `List`
   * Replace `VaultManagerActivity` with `VaultManagerFragment`
+  * Remove `clientToken` and `tokenizationKey` from `DropInRequest` in favor of instantiating `DropInClient` with authorization
 
 **Note:** The credentials for integrating with 3DS have changed. If you are using 3DS please update the credentials in your app-level `build.gradle`